Implement FavoriteServiceImpl
