+++ itachallenge-challenge/src/main/java/com/itachallenge/challenge/service/ChallengeServiceImpl.java
@@ -56,6 +56,8 @@ public class ChallengeServiceImpl implements IChallengeService {
     private IUserService userService;
     @Autowired
     private ITagService tagService;
+    @Autowired
+    private IFavoriteService favoriteService;
 
     @Cacheable(value = "challenges", key = "#id", unless = "#result==null")
     public Mono<ChallengeDto> getChallengeById(String id) {
@@ -351,55 +353,12 @@ public class ChallengeServiceImpl implements IChallengeService {
 
     @Override
     public Mono<FavoriteDto> addChallengeToFavorites(String challengeId, String userId) {
-
-        Mono<UUID> challengeIdMono = validateUUID(String.valueOf(challengeId));
-        Mono<UUID> userIdMono = validateUUID(String.valueOf(userId));
-
-        return Mono.zip(challengeIdMono, userIdMono)
-                .flatMap(Uuidtuple -> {
-                    UUID challengeUuid = Uuidtuple.getT1();
-                    UUID userUuid = Uuidtuple.getT2();
-
-                    return challengeRepository.findByUuid(challengeUuid)
-                            .switchIfEmpty(Mono.error(new ChallengeNotFoundException(String.format(CHALLENGE_NOT_FOUND_ERROR, challengeUuid))))
-                            .flatMap(challenge -> userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())
-                                    .onErrorResume(throwable -> Mono.error(new InternalServerErrorException(throwable.getMessage())))
-                                    .flatMap(isAddedToUsersFavorites -> {
-                                        if (Boolean.TRUE.equals(isAddedToUsersFavorites) ||
-                                                Optional.ofNullable(challenge.getTimesFavorite()).orElse(0) == 0) {
-                                            challenge.increaseTimesFavorite();
-                                            return challengeRepository.save(challenge);
-                                        }
-                                        return Mono.just(challenge);
-                                    })
-                                    .map(savedChallenge -> new FavoriteDto(true, savedChallenge.getTimesFavorite())));
-                });
+        return favoriteService.addChallengeToFavorites(challengeId, userId);
     }
 
     @Override
     public Mono<FavoriteDto> removeChallengeFromFavorites(String challengeId, String userId) {
-        Mono<UUID> challengeIdMono = validateUUID(String.valueOf(challengeId));
-        Mono<UUID> languageIdMono = validateUUID(String.valueOf(userId));
-
-        return Mono.zip(challengeIdMono, languageIdMono)
-                .flatMap(Uuidtuple -> {
-                    UUID challengeUuid = Uuidtuple.getT1();
-                    UUID userUuid = Uuidtuple.getT2();
-
-                    return challengeRepository.findByUuid(challengeUuid)
-                            .switchIfEmpty(Mono.error(new ChallengeNotFoundException(String.format(CHALLENGE_NOT_FOUND_ERROR, challengeUuid))))
-                            .flatMap(challenge -> userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())
-                                    .onErrorResume(throwable -> Mono.error(new InternalServerErrorException(throwable.getMessage())))
-                                    .flatMap(isRemovedFromUsersFavorites -> {
-                                        if (Boolean.TRUE.equals(isRemovedFromUsersFavorites) ||
-                                                Optional.ofNullable(challenge.getTimesFavorite()).orElse(0) == 0) {
-                                            challenge.decreaseTimesFavorite();
-                                            return challengeRepository.save(challenge);
-                                        }
-                                        return Mono.just(challenge);
-                                    })
-                                    .map(savedChallenge -> new FavoriteDto(false, savedChallenge.getTimesFavorite())));
-                });
+        return favoriteService.removeChallengeFromFavorites(challengeId, userId);
     }
 
     @Override
