+++ itachallenge-challenge/src/test/java/com/itachallenge/challenge/service/ChallengeServiceImplTest.java
@@ -658,15 +658,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
     }
 
     @Test
-    void addChallengeToFavorites_WhenChallengeUuidNotValid_ReturnsError() {
-        StepVerifier.create(challengeService.addChallengeToFavorites("InvalidUuid", UUID.randomUUID().toString()))
-                .expectErrorMatches(error ->
-                        error instanceof BadUUIDException &&
-                                error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
-                .verify();
-    }
-
-    @Test
     void addChallengeToBookmarks_WhenChallengeUuidNotValid_ReturnsError() {
         StepVerifier.create(challengeService.addChallengeToBookmarks("InvalidUuid", UUID.randomUUID().toString()))
                 .expectErrorMatches(error ->
@@ -678,15 +669,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
     @Test
     void addChallengeToSolved_WhenChallengeUuidNotValid_ReturnsError() {
         StepVerifier.create(challengeService.addChallengeToSolved("InvalidUuid", UUID.randomUUID().toString()))
-                .expectErrorMatches(error ->
-                        error instanceof BadUUIDException &&
-                                error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
-                .verify();
-    }
-
-    @Test
-    void addChallengeToFavorites_WhenUserUuidNotValid_ReturnsError() {
-        StepVerifier.create(challengeService.addChallengeToFavorites(UUID.randomUUID().toString(), "InvalidUuid"))
                 .expectErrorMatches(error ->
                         error instanceof BadUUIDException &&
                                 error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
@@ -712,15 +694,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
     }
 
     @Test
-    void addChallengeToFavorites_WhenChallengeUuidIsNull_ReturnsError() {
-        StepVerifier.create(challengeService.addChallengeToFavorites(null, UUID.randomUUID().toString()))
-                .expectErrorMatches(error ->
-                        error instanceof BadUUIDException &&
-                                error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
-                .verify();
-    }
-
-    @Test
     void addChallengeToBookmarks_WhenChallengeUuidIsNull_ReturnsError() {
         StepVerifier.create(challengeService.addChallengeToBookmarks(null, UUID.randomUUID().toString()))
                 .expectErrorMatches(error ->
@@ -732,15 +705,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
     @Test
     void addChallengeToSolved_WhenChallengeUuidIsNull_ReturnsError() {
         StepVerifier.create(challengeService.addChallengeToSolved(null, UUID.randomUUID().toString()))
-                .expectErrorMatches(error ->
-                        error instanceof BadUUIDException &&
-                                error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
-                .verify();
-    }
-
-    @Test
-    void addChallengeToFavorites_WhenUserUuidIsNull_ReturnsError() {
-        StepVerifier.create(challengeService.addChallengeToFavorites(UUID.randomUUID().toString(), null))
                 .expectErrorMatches(error ->
                         error instanceof BadUUIDException &&
                                 error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
@@ -763,22 +727,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
                         error instanceof BadUUIDException &&
                                 error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
                 .verify();
-    }
-
-    @Test
-    void addChallengeToFavorites_WhenChallengeNotFound_ReturnsError() {
-        String CHALLENGE_NOT_FOUND_ERROR = "Challenge with id: %s not found";
-        UUID challengeUuid = UUID.randomUUID();
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.empty());
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), UUID.randomUUID().toString()))
-                .expectErrorMatches(error ->
-                        error instanceof ChallengeNotFoundException &&
-                                error.getMessage().equals(String.format(CHALLENGE_NOT_FOUND_ERROR, challengeUuid.toString())))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
     }
 
     @Test
@@ -814,27 +762,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
     }
 
     @Test
-    void addChallengeToFavorites_WhenUserNotFound_ReturnsError() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        String message = "Some error message";
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-
-        when(userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.error(new UserNotFoundException(message)));
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectErrorMatches(error ->
-                        error instanceof InternalServerErrorException &&
-                                error.getMessage().equals(message))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).addChallengeToFavorites(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
     void addChallengeToBookmarks_WhenUserNotFound_ReturnsError() {
         UUID challengeUuid = UUID.randomUUID();
         UUID userUuid = UUID.randomUUID();
@@ -853,27 +780,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
 
         verify(challengeRepository, times(1)).findByUuid(challengeUuid);
         verify(userService, times(1)).addChallengeToBookmarks(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
-    void addChallengeToFavorites_WhenUserServiceReturnsCustomBadRequestException_ReturnsError() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        String message = "Some error message";
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-
-        when(userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.error(new BadRequestException(message)));
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectErrorMatches(error ->
-                        error instanceof InternalServerErrorException &&
-                                error.getMessage().equals(message))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).addChallengeToFavorites(userUuid.toString(), challengeUuid.toString());
     }
 
     @Test
@@ -898,27 +804,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
     }
 
     @Test
-    void addChallengeToFavorites_WhenUserServiceReturnsCustomInternalServerErrorException_ReturnsError() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        String message = "Some error message";
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-
-        when(userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.error(new InternalServerErrorException(message)));
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectErrorMatches(error ->
-                        error instanceof InternalServerErrorException &&
-                                error.getMessage().equals(message))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).addChallengeToFavorites(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
     void addChallengeToBookmarks_WhenUserServiceReturnsCustomInternalServerErrorException_ReturnsError() {
         UUID challengeUuid = UUID.randomUUID();
         UUID userUuid = UUID.randomUUID();
@@ -940,27 +825,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
     }
 
     @Test
-    void addChallengeToFavorites_WhenUserServiceReturnsAnyException_ReturnsError() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        String message = "Some error message";
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-
-        when(userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.error(new Exception(message)));
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectErrorMatches(error ->
-                        error instanceof Exception &&
-                                error.getMessage().equals(message))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).addChallengeToFavorites(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
     void addChallengeToBookmarks_WhenUserServiceReturnsAnyException_ReturnsError() {
         UUID challengeUuid = UUID.randomUUID();
         UUID userUuid = UUID.randomUUID();
@@ -979,33 +843,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
 
         verify(challengeRepository, times(1)).findByUuid(challengeUuid);
         verify(userService, times(1)).addChallengeToBookmarks(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
-    void addChallengeToFavorites_WhenAdded_IncreasesTimesFavoriteAndReturnsFavoriteDTO() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        int initialTimesFavorite = 20;
-
-        challenge.setTimesFavorite(initialTimesFavorite);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(true));
-        when(challengeRepository.save(challenge)).thenReturn(Mono.just(challenge));
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                        return favoriteDto.getTimesFavorited() == initialTimesFavorite + 1 &&
-                                favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(initialTimesFavorite + 1, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).addChallengeToFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(1)).save(challenge);
     }
 
     @Test
@@ -1036,32 +873,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
     }
 
     @Test
-    void addChallengeToFavorites_WhenAddedAndInitialTimesFavoriteIsNull_IncreasesTimesFavoriteAndReturnsFavoriteDTO() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-
-        challenge.setTimesFavorite(null);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(true));
-        when(challengeRepository.save(challenge)).thenReturn(Mono.just(challenge));
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                    return favoriteDto.getTimesFavorited() == 1 &&
-                            favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(1, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).addChallengeToFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(1)).save(challenge);
-    }
-
-    @Test
     void addChallengeToBookmarks_WhenAddedAndInitialTimesFavoriteIsNull_IncreasesTimesBookmarkAndReturnsBookmarkDTO() {
         UUID challengeUuid = UUID.randomUUID();
         UUID userUuid = UUID.randomUUID();
@@ -1085,32 +896,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
         verify(challengeRepository, times(1)).findByUuid(challengeUuid);
         verify(userService, times(1)).addChallengeToBookmarks(userUuid.toString(), challengeUuid.toString());
         verify(challengeRepository, times(1)).save(challenge);
-    }
-
-    @Test
-    void addChallengeToFavorites_WhenNotAdded_NotIncreaseTimesFavoriteAndReturnsFavoriteDTO() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        int initialTimesFavorite = 20;
-
-        challenge.setTimesFavorite(initialTimesFavorite);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(false));
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                    return favoriteDto.getTimesFavorited() == initialTimesFavorite &&
-                            favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(initialTimesFavorite, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).addChallengeToFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(0)).save(any());
     }
 
     @Test
@@ -1141,33 +926,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
 
     @ParameterizedTest
     @MethodSource
-    void addChallengeToFavorites_WhenNotAddedAndTimesFavoriteIsNullOrZero_SetTimesFavoriteToOneAndReturnsFavoriteDTO(Integer timesFavorite) {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-
-        challenge.setTimesFavorite(timesFavorite);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.addChallengeToFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(false));
-        when(challengeRepository.save(challenge)).thenReturn(Mono.just(challenge));
-
-        StepVerifier.create(challengeService.addChallengeToFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                    return favoriteDto.getTimesFavorited() == 1 &&
-                            favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(1, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).addChallengeToFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(1)).save(challenge);
-    }
-
-    @ParameterizedTest
-    @MethodSource
     void addChallengeToBookmarks_WhenNotAddedAndTimesBookmarkIsNullOrZero_SetTimesBookmarkToOneAndReturnsBookmarkDTO(Integer timesBookmark) {
         UUID challengeUuid = UUID.randomUUID();
         UUID userUuid = UUID.randomUUID();
@@ -1193,284 +951,12 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
         verify(challengeRepository, times(1)).save(challenge);
     }
 
-    public static Stream<Integer> addChallengeToFavorites_WhenNotAddedAndTimesFavoriteIsNullOrZero_SetTimesFavoriteToOneAndReturnsFavoriteDTO() {
-        return Stream.of(null, 0);
-    }
-
     public static Stream<Integer> addChallengeToBookmarks_WhenNotAddedAndTimesBookmarkIsNullOrZero_SetTimesBookmarkToOneAndReturnsBookmarkDTO() {
         return Stream.of(null, 0);
     }
 
     public static Stream<Integer> addChallengeToSolved_WhenNotAddedAndTimesSolvedIsNullOrZero_SetTimesSolvedToOneAndReturnsSolvedDTO() {
         return Stream.of(null, 0);
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenChallengeUuidNotValid_ReturnsError() {
-        StepVerifier.create(challengeService.removeChallengeFromFavorites("InvalidUuid", UUID.randomUUID().toString()))
-                .expectErrorMatches(error ->
-                        error instanceof BadUUIDException &&
-                                error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
-                .verify();
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenUserUuidNotValid_ReturnsError() {
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(UUID.randomUUID().toString(), "InvalidUuid"))
-                .expectErrorMatches(error ->
-                        error instanceof BadUUIDException &&
-                                error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
-                .verify();
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenChallengeUuidIsNull_ReturnsError() {
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(null, UUID.randomUUID().toString()))
-                .expectErrorMatches(error ->
-                        error instanceof BadUUIDException &&
-                                error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
-                .verify();
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenUserUuidIsNull_ReturnsError() {
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(UUID.randomUUID().toString(), null))
-                .expectErrorMatches(error ->
-                        error instanceof BadUUIDException &&
-                                error.getMessage().equals("Invalid ID format. Please indicate the correct format."))
-                .verify();
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenChallengeNotFound_ReturnsError() {
-        String CHALLENGE_NOT_FOUND_ERROR = "Challenge with id: %s not found";
-        UUID challengeUuid = UUID.randomUUID();
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.empty());
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), UUID.randomUUID().toString()))
-                .expectErrorMatches(error ->
-                        error instanceof ChallengeNotFoundException &&
-                                error.getMessage().equals(String.format(CHALLENGE_NOT_FOUND_ERROR, challengeUuid.toString())))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenUserNotFound_ReturnsError() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        String message = "Some error message";
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.error(new UserNotFoundException(message)));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectErrorMatches(error ->
-                        error instanceof InternalServerErrorException &&
-                                error.getMessage().equals(message))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenUserServiceReturnsCustomBadRequestException_ReturnsError() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        String message = "Some error message";
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.error(new BadRequestException(message)));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectErrorMatches(error ->
-                        error instanceof InternalServerErrorException &&
-                                error.getMessage().equals(message))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenUserServiceReturnsCustomInternalServerErrorException_ReturnsError() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        String message = "Some error message";
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.error(new InternalServerErrorException(message)));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectErrorMatches(error ->
-                        error instanceof InternalServerErrorException &&
-                                error.getMessage().equals(message))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenUserServiceReturnsAnyException_ReturnsError() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        String message = "Some error message";
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.error(new Exception(message)));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectErrorMatches(error ->
-                        error instanceof Exception &&
-                                error.getMessage().equals(message))
-                .verify();
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenRemoved_DecreasesTimesFavoriteAndReturnsFavoriteDTO() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        int initialTimesFavorite = 20;
-
-        challenge.setTimesFavorite(initialTimesFavorite);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(true));
-        when(challengeRepository.save(challenge)).thenReturn(Mono.just(challenge));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                    return favoriteDto.getTimesFavorited() == initialTimesFavorite - 1 &&
-                            !favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(initialTimesFavorite - 1, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(1)).save(challenge);
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenRemovedAndInitialTimesFavoriteIsNull_SetsTimesFavoriteToZeroAndReturnsFavoriteDTO() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-
-        challenge.setTimesFavorite(null);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(true));
-        when(challengeRepository.save(challenge)).thenReturn(Mono.just(challenge));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                    return favoriteDto.getTimesFavorited() == 0 &&
-                            !favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(0, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(1)).save(challenge);
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenRemovedAndInitialTimesFavoriteIsZero_SetsTimesFavoriteToZeroAndReturnsFavoriteDTO() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-
-        challenge.setTimesFavorite(0);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(true));
-        when(challengeRepository.save(challenge)).thenReturn(Mono.just(challenge));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                    return favoriteDto.getTimesFavorited() == 0 &&
-                            !favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(0, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(1)).save(challenge);
-    }
-
-    @Test
-    void removeChallengeFromFavorites_WhenNotRemoved_NotChangeTimesFavoriteAndReturnsFavoriteDTO() {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-        int initialTimesFavorite = 20;
-
-        challenge.setTimesFavorite(initialTimesFavorite);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(false));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                    return favoriteDto.getTimesFavorited() == initialTimesFavorite &&
-                            !favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(initialTimesFavorite, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(0)).save(any());
-    }
-
-    @ParameterizedTest
-    @MethodSource
-    void removeChallengeFromFavorites_WhenNotRemovedAndTimesFavoriteIsNullOrZero_SetTimesFavoriteToZeroAndReturnsFavoriteDTO(Integer timesFavorite) {
-        UUID challengeUuid = UUID.randomUUID();
-        UUID userUuid = UUID.randomUUID();
-        ChallengeDocument challenge = new ChallengeDocument();
-
-        challenge.setTimesFavorite(timesFavorite);
-
-        when(challengeRepository.findByUuid(challengeUuid)).thenReturn(Mono.just(challenge));
-        when(userService.removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString())).thenReturn(Mono.just(false));
-        when(challengeRepository.save(challenge)).thenReturn(Mono.just(challenge));
-
-        StepVerifier.create(challengeService.removeChallengeFromFavorites(challengeUuid.toString(), userUuid.toString()))
-                .expectNextMatches(favoriteDto -> {
-                    return favoriteDto.getTimesFavorited() == 0 &&
-                            !favoriteDto.isFavorite();
-                })
-                .verifyComplete();
-
-        Assertions.assertEquals(0, challenge.getTimesFavorite());
-
-        verify(challengeRepository, times(1)).findByUuid(challengeUuid);
-        verify(userService, times(1)).removeChallengeFromFavorites(userUuid.toString(), challengeUuid.toString());
-        verify(challengeRepository, times(1)).save(challenge);
     }
 
     @Test
@@ -1639,11 +1125,6 @@ void addChallengeToSolved_WhenChallengeAlreadySolved_DoesNotIncreaseTimesSolvedA
                     assertEquals(0, genericResultDto.getCount());
                     return true;
                 });
-    }
-
-
-    public static Stream<Integer> removeChallengeFromFavorites_WhenNotRemovedAndTimesFavoriteIsNullOrZero_SetTimesFavoriteToZeroAndReturnsFavoriteDTO() {
-        return Stream.of(null, 0);
     }
 
     @Test
